Format a ProTracker pattern conform OpenMPT specs

OpenMpt is a popular modern music tracker. This function allows you to format a pattern such that it can be pasted directly into OpenMPT. On Windows you can use writeClipboard() for this purpose.

Usage

as_modplug_pattern(pattern, ...)

Arguments

pattern

An object of class pt2pat to be formatted

...

Ignored

Value

Returns a character object formatted such that it can be copied into OpenMPT

Author

Pepijn de Vries

Examples

mod    <- pt2_read_mod(pt2_demo())
as_modplug_pattern(pt2_pattern(mod, 0L))
#>  [1] "ModPlug Tracker MOD"                             
#>  [2] "|C-301...A08|C-302...C40|A#103...F06|A#204......"
#>  [3] "|...01...A08|C-302...C30|........A01|........A02"
#>  [4] "|...01...A08|C-302......|........A01|C-304......"
#>  [5] "|...........|C-302......|........A02|........A02"
#>  [6] "|...01...A08|C-302......|........120|D-304......"
#>  [7] "|...........|A#204...C08|........A02|........A02"
#>  [8] "|...01...A08|C-302......|........220|D#304......"
#>  [9] "|...........|A#204...C08|........A01|........A02"
#> [10] "|...01...A08|C-302......|........A01|F-304......"
#> [11] "|...........|A#304...C08|........A01|........A02"
#> [12] "|...01...A08|C-302......|........A01|D#304......"
#> [13] "|...........|G-304...C08|........A01|........A02"
#> [14] "|G-301......|C-302......|........A01|D-304......"
#> [15] "|........A08|F-304...C08|...........|........A02"
#> [16] "|F-301......|C-302......|...........|C-304......"
#> [17] "|........A08|A#204...C08|...........|........A02"
#> [18] "|C-301...A08|C-302...C40|A#103...F06|A#204......"
#> [19] "|...01...A08|C-302...C30|........A01|........A02"
#> [20] "|...01...A08|C-302......|........A01|D-304......"
#> [21] "|...........|C-302......|........A02|........A02"
#> [22] "|...01...A08|C-302......|........120|F-204......"
#> [23] "|...........|A#204...C08|........A02|........A02"
#> [24] "|...01...A08|C-302......|........220|G-204......"
#> [25] "|...........|A#204...C08|........A01|........A02"
#> [26] "|...01...A08|C-302......|........A01|A#204......"
#> [27] "|...........|A#304...C08|........A01|........A01"
#> [28] "|...01...A08|C-302......|........A01|..........."
#> [29] "|...........|G-304...C08|........A01|........A01"
#> [30] "|G-201......|C-302......|........A01|..........."
#> [31] "|........A08|F-204...C08|...........|........A01"
#> [32] "|A-201......|C-302......|...........|..........."
#> [33] "|........A08|G-204...C08|...........|........A01"
#> [34] "|E-301...A08|C-302...C40|D-203......|D-304......"
#> [35] "|...01...A08|C-302...C30|........A01|........A02"
#> [36] "|...01...A08|C-302......|........A01|E-304......"
#> [37] "|...........|C-302......|........A02|........A02"
#> [38] "|...01...A08|C-302......|........126|F#304......"
#> [39] "|...........|D-304...C08|........A02|........A02"
#> [40] "|...01...A08|C-302......|........226|G-304......"
#> [41] "|...........|E-304...C08|........A01|........A02"
#> [42] "|...01...A08|C-302......|........A01|A-304......"
#> [43] "|...........|D-304...C08|........A01|........A02"
#> [44] "|...01...A08|C-302......|........A01|G-304......"
#> [45] "|...........|D-304...C08|........A01|........A02"
#> [46] "|B-301......|C-302......|........A01|F#304......"
#> [47] "|........A08|D-304...C08|...........|........A02"
#> [48] "|A-301......|C-302......|...........|E-304......"
#> [49] "|........A08|E-204...C08|...........|........A02"
#> [50] "|D-301...A08|C-302...C40|C-203......|C-304......"
#> [51] "|...01...A08|C-302...C30|........A01|........A02"
#> [52] "|...01...A08|C-302......|........A01|D-304......"
#> [53] "|...........|C-302......|........A02|........A02"
#> [54] "|...01...A08|C-302......|........12B|E-304......"
#> [55] "|...........|G-304...C08|........A02|........A02"
#> [56] "|...01...A08|C-302......|........22B|F-304......"
#> [57] "|...........|G-304...C08|........A01|........A02"
#> [58] "|...01...A08|C-302......|........A01|G-304......"
#> [59] "|...........|E-304...C08|........A01|........A02"
#> [60] "|...01...A08|C-302......|........A01|F-304......"
#> [61] "|...........|C-304...C08|........A01|........A02"
#> [62] "|A-301......|C-302......|........A01|E-304......"
#> [63] "|........A08|G-304...C08|...........|........A02"
#> [64] "|G-301......|F-304...C08|...........|D-304......"
#> [65] "|........A08|C-304...C08|...........|........A02"