Retrieve a pattern from a ProTracker module

Get a pattern table (sequence of notes and effects on each of the 4 channels) at a specific index from a ProTracker module.

Usage

pt2_pattern(mod, i, ...)

Arguments

mod

A pt2mod class objects from which to retrieve a pattern table

i

The index (integer) of the pattern. Note that the index starts at 0.

...

Ignored

Value

A pt2pat object representing the pattern.

Author

Pepijn de Vries

Examples

mod <- pt2_read_mod(pt2_demo())

pt2_pattern(mod, 0L)
#> rw  chnnl 1     chnnl 2     chnnl 3     chnnl 4   
#> 00  C-3 01 A08  C-3 02 C40  A#1 03 F06  A#2 04 000
#> 01  --- 01 A08  C-3 02 C30  --- 00 A01  --- 00 A02
#> 02  --- 01 A08  C-3 02 000  --- 00 A01  C-3 04 000
#> 03  --- 00 000  C-3 02 000  --- 00 A02  --- 00 A02
#> 04  --- 01 A08  C-3 02 000  --- 00 120  D-3 04 000
#> 05  --- 00 000  A#2 04 C08  --- 00 A02  --- 00 A02
#> 06  --- 01 A08  C-3 02 000  --- 00 220  D#3 04 000
#> 07  --- 00 000  A#2 04 C08  --- 00 A01  --- 00 A02
#> 08  --- 01 A08  C-3 02 000  --- 00 A01  F-3 04 000
#> 09  --- 00 000  A#3 04 C08  --- 00 A01  --- 00 A02
#> 10  --- 01 A08  C-3 02 000  --- 00 A01  D#3 04 000
#> 11  --- 00 000  G-3 04 C08  --- 00 A01  --- 00 A02
#> 12  G-3 01 000  C-3 02 000  --- 00 A01  D-3 04 000
#> 13  --- 00 A08  F-3 04 C08  --- 00 000  --- 00 A02
#> 14  F-3 01 000  C-3 02 000  --- 00 000  C-3 04 000
#> 15  --- 00 A08  A#2 04 C08  --- 00 000  --- 00 A02
#> 16  C-3 01 A08  C-3 02 C40  A#1 03 F06  A#2 04 000
#> 17  --- 01 A08  C-3 02 C30  --- 00 A01  --- 00 A02
#> 18  --- 01 A08  C-3 02 000  --- 00 A01  D-3 04 000
#> 19  --- 00 000  C-3 02 000  --- 00 A02  --- 00 A02
#> 20  --- 01 A08  C-3 02 000  --- 00 120  F-2 04 000
#> 21  --- 00 000  A#2 04 C08  --- 00 A02  --- 00 A02
#> 22  --- 01 A08  C-3 02 000  --- 00 220  G-2 04 000
#> 23  --- 00 000  A#2 04 C08  --- 00 A01  --- 00 A02
#> 24  --- 01 A08  C-3 02 000  --- 00 A01  A#2 04 000
#> 25  --- 00 000  A#3 04 C08  --- 00 A01  --- 00 A01
#> 26  --- 01 A08  C-3 02 000  --- 00 A01  --- 00 000
#> 27  --- 00 000  G-3 04 C08  --- 00 A01  --- 00 A01
#> 28  G-2 01 000  C-3 02 000  --- 00 A01  --- 00 000
#> 29  --- 00 A08  F-2 04 C08  --- 00 000  --- 00 A01
#> 30  A-2 01 000  C-3 02 000  --- 00 000  --- 00 000
#> 31  --- 00 A08  G-2 04 C08  --- 00 000  --- 00 A01
#> 32  E-3 01 A08  C-3 02 C40  D-2 03 000  D-3 04 000
#> 33  --- 01 A08  C-3 02 C30  --- 00 A01  --- 00 A02
#> 34  --- 01 A08  C-3 02 000  --- 00 A01  E-3 04 000
#> 35  --- 00 000  C-3 02 000  --- 00 A02  --- 00 A02
#> 36  --- 01 A08  C-3 02 000  --- 00 126  F#3 04 000
#> 37  --- 00 000  D-3 04 C08  --- 00 A02  --- 00 A02
#> 38  --- 01 A08  C-3 02 000  --- 00 226  G-3 04 000
#> 39  --- 00 000  E-3 04 C08  --- 00 A01  --- 00 A02
#> 40  --- 01 A08  C-3 02 000  --- 00 A01  A-3 04 000
#> 41  --- 00 000  D-3 04 C08  --- 00 A01  --- 00 A02
#> 42  --- 01 A08  C-3 02 000  --- 00 A01  G-3 04 000
#> 43  --- 00 000  D-3 04 C08  --- 00 A01  --- 00 A02
#> 44  B-3 01 000  C-3 02 000  --- 00 A01  F#3 04 000
#> 45  --- 00 A08  D-3 04 C08  --- 00 000  --- 00 A02
#> 46  A-3 01 000  C-3 02 000  --- 00 000  E-3 04 000
#> 47  --- 00 A08  E-2 04 C08  --- 00 000  --- 00 A02
#> 48  D-3 01 A08  C-3 02 C40  C-2 03 000  C-3 04 000
#> 49  --- 01 A08  C-3 02 C30  --- 00 A01  --- 00 A02
#> 50  --- 01 A08  C-3 02 000  --- 00 A01  D-3 04 000
#> 51  --- 00 000  C-3 02 000  --- 00 A02  --- 00 A02
#> 52  --- 01 A08  C-3 02 000  --- 00 12B  E-3 04 000
#> 53  --- 00 000  G-3 04 C08  --- 00 A02  --- 00 A02
#> 54  --- 01 A08  C-3 02 000  --- 00 22B  F-3 04 000
#> 55  --- 00 000  G-3 04 C08  --- 00 A01  --- 00 A02
#> 56  --- 01 A08  C-3 02 000  --- 00 A01  G-3 04 000
#> 57  --- 00 000  E-3 04 C08  --- 00 A01  --- 00 A02
#> 58  --- 01 A08  C-3 02 000  --- 00 A01  F-3 04 000
#> 59  --- 00 000  C-3 04 C08  --- 00 A01  --- 00 A02
#> 60  A-3 01 000  C-3 02 000  --- 00 A01  E-3 04 000
#> 61  --- 00 A08  G-3 04 C08  --- 00 000  --- 00 A02
#> 62  G-3 01 000  F-3 04 C08  --- 00 000  D-3 04 000
#> 63  --- 00 A08  C-3 04 C08  --- 00 000  --- 00 A02